feat: program stack is allocated using mm

+++ x64barebones/Kernel/include/memoryManager.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-typedef size_t header_t;
+typedef uint64_t header_t;
 
 #define BITS_IN_BYTE 8
 
@@ -12,7 +12,7 @@ typedef size_t header_t;
 #define TRUE 1
 #define FALSE (!TRUE)
 
-#define HEAP_SIZE 0x100000 // 1Mb (in bytes!!!!)    
+#define HEAP_SIZE 0x100000  // 1Mb (in bytes!!!!)    
 #define HEAP_START (header_t *) 0xA00000  // 10 Mb  (TODO: habria que calcular el fin del userland y meterlo ahi)
 #define HEAP_END (HEAP_START + HEAP_SIZE / BITS_IN_BYTE)
